fix(ingest): ignore dataProcessInstance from stateful ingestion
